… Cargo.lock
… crates/matrix-sdk-ui/Cargo.toml
… crates/matrix-sdk-ui/src/timeline/pinned_events_loader.rs
… crates/matrix-sdk/src/event_cache/mod.rs
… crates/matrix-sdk/src/event_cache/pagination.rs
… crates/matrix-sdk/src/event_cache/paginator.rs
… crates/matrix-sdk/src/room/edit.rs