This series of patches, investigated using Fasterthanlime's [good advice](https://fasterthanli.me/articles/why-is-my-rust-build-so-slow#rustc-self-profiling), helped lowering the compiles times from 44 seconds to 4 on my machine, for the `matrix-sdk-ui` crate.

Compile times for the `matrix-sdk` crate have also been lowered from 33 to 12 seconds.

Overall, for a plain `cargo build`:

- before: 113 seconds
- after: 54 seconds

Each commit can be reviewed independently.

See also: https://github.com/rust-lang/rust/issues/87012